components: add logo component and refactor
