+++ src/components/LoginBrandInfo/styles.module.scss
@@ -20,13 +20,10 @@
 		width: 100%;
 		margin-top: 25rem;
 
-		.logoWrapper {
-			.logo {
-				width: 10rem;
-				height: 3rem;
-				cursor: pointer;
-				margin: 1rem auto;
-			}
+		.logo {
+			width: 10rem;
+			height: 3rem;
+			margin: 1rem auto;
 		}
 
 		p {
