+++ src/components/Logo/styles.module.scss
@@ -1,0 +1,6 @@
+.logo {
+	margin: 2rem auto;
+	width: 9rem;
+	height: 3rem;
+	cursor: pointer;
+}
