+++ src/components/Signup/index.tsx
@@ -9,13 +9,14 @@ import BarLoader from 'react-spinners/ClipLoader';
 import Button from 'components/common/Button';
 import ErrorMessage from 'components/common/Messages/ErrorMessage';
 import Image from 'components/common/Image';
-import Input from 'components/common/Inputs/Input';
-import PasswordInput from 'components/common/Inputs/PasswordInput';
+import Input from 'components/common/Input';
+import PasswordInput from 'components/common/Input/PasswordInput';
 
 import googleIcon from 'assets/icons/google.svg';
 import logoIcon from 'assets/icons/logo.svg';
 
 import styles from './styles.module.scss';
+import Logo from 'components/Logo';
 
 YupPassword(Yup);
 
@@ -67,18 +68,7 @@ const SignupComponent: FC<Props> = ({ isMobile }) => {
 
 	return (
 		<div className={styles.signUpContainer}>
-			{isMobile && (
-				<Link href='/' passHref>
-					<a className={styles.logoWrapper}>
-						<Image
-							src={logoIcon}
-							alt='people-logo'
-							className={styles.logo}
-							lazy={false}
-						/>
-					</a>
-				</Link>
-			)}
+			{isMobile && <Logo className={styles.logo} />}
 			<div className={styles.heading}>
 				<h1>Create new account</h1>
 				<p>Please fill in the form to continue</p>
