+++ src/components/Signup/styles.module.scss
@@ -26,14 +26,8 @@
 		padding: 2rem 2rem;
 	}
 
-	.logoWrapper {
-		margin-bottom: 4rem;
-
-		.logo {
-			width: 9rem;
-			height: 3rem;
-			cursor: pointer;
-		}
+	.logo {
+		margin: 0 auto 4rem auto;
 	}
 
 	.heading {
